Moves the editor and sscanss module into a single source directory and moves shared dependency into core module
