+++ sscanss/app/widgets/__init__.py
@@ -1,0 +1,2 @@
+from .graphics import GraphicsView, GraphicsScene, GraphicsPointItem, Grid
+from .table_model import PointModel, AlignmentErrorModel, ErrorDetailModel, CenteredBoxProxy, LimitTextDelegate
