+++ sscanss/core/io/__init__.py
@@ -1,3 +1,4 @@
 from .reader import (read_3d_model, read_obj, read_stl, read_project_hdf, read_points, read_vectors, read_trans_matrix,
-                     read_fpos, validate_vector_length, read_kinematic_calibration_file, read_robot_world_calibration_file)
+                     read_fpos, validate_vector_length, read_kinematic_calibration_file,
+                     read_robot_world_calibration_file)
 from .writer import write_project_hdf, write_binary_stl, write_points
