+++ sscanss/core/scene/__init__.py
@@ -4,3 +4,4 @@ from .entity import (SampleEntity, FiducialEntity, MeasurementPointEntity, Measu
 from .camera import Camera, world_to_screen, screen_to_world
 from .scene import Scene, validate_instrument_scene_size
 from .shader import VertexArray, GouraudShader, DefaultShader, Shader
+from .renderer import OpenGLRenderer
