+++ sscanss/core/util/__init__.py
@@ -1,3 +1,7 @@
 from .misc import (Directions, Primitives, to_float, TransformType, DockFlag, PointType, Attributes, POINT_DTYPE,
                    StrainComponents, LoadVector, MessageSeverity, CommandID, toggleActionInGroup, PlaneOptions)
 from .worker import Worker
+from .forms import FormGroup, FormControl, CompareValidator, Banner, FormTitle
+from .widgets import (create_icon, create_header, create_scroll_area, create_tool_button, FileDialog, FilePicker,
+                      Accordion, Pane, StatusBar, ColourPicker)
+
