+++ sscanss/editor/scene_manager.py
