+++ sscanss/editor/widgets.py
